# Enhance query performance