--- a/scripts/stations_database_setup.py
+++ b/scripts/stations_database_setup.py
@@ -136,7 +136,7 @@ def assign_freq_band_count():
 def get_latitude_segment(latitude):
     """Returns the segment index for a given latitude."""
     base_latitude = 49.0  # Starting latitude
-    segment_size = 0.3  # Size of each latitude segment // 33.(3) km
+    segment_size = 0.1  # Size of each latitude segment // 11.(1) km
     return int((latitude - base_latitude) / segment_size)
 
 def assign_latitude_segments():