This commit introduces significant optimizations to the spatial querying mechanism resulting in powerful performance improvements.

Key Changes:
- Increased db dataset size from 21 to 65 segments, significantly enhancing the speed of data processing thus reducing time delays and improving user response times

Previous Improvements:

#116 #108

Performance Improvements:

I. in local env:

- Without filters: response times were reduced from ~0.5-1s to ~0.120s / ~**84%** reduction in latency
- With filters applied: from ~0.1s to ~0.042s / ~**58%** reduction in latency

II. in PROD env:

- Without filters: from ~0.5-0.7s to ~TBC seconds
- With filters: from ~0.2s to ~TBC seconds